pyproject.toml
[project]
name = "mistral-reddit-sentiment"
version = "0.1.0"
description = "Weekly sentiment analysis for r/MistralAI using Reddit and selectable LLM providers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.52.0",
  "crawl4ai>=0.8.0",
  "discord.py>=2.4.0",
  "fastapi>=0.116.0",
  "fastmcp>=2.13.0",
  "gspread>=6.1.2",
  "python-dotenv>=1.1.0",
  "requests>=2.32.0",
  "uvicorn>=0.35.0",
]

[project.scripts]
mistral-sentiment = "mistral_sentiment_app.main:main"
mistral-sentiment-api = "mistral_sentiment_app.api:run"
mistral-sentiment-mcp = "mistral_sentiment_app.mcp_server:run"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/mistral_sentiment_app"]
